Rename federation registration endpoint
